Refactor: Profile and AnimeBacklog Refactored AnimeBacklog use JoinTable annotations to accurately build the desired database schema and also dropped the name field. Added cascading to support updating underlying collections. Disabled CSRF filtering (for the time being). Changed dbcp driver.

+++ src/main/java/com/nullteam6/utility/PaginatedList.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class PaginatedList<T> {
-    private int totalCount;
+    private long totalCount;
     private String next;
     private String last;
     private List<T> data;
@@ -13,11 +13,11 @@ public class PaginatedList<T> {
         data = new ArrayList<>();
     }
 
-    public int getTotalCount() {
+    public long getTotalCount() {
         return totalCount;
     }
 
-    public void setTotalCount(int totalCount) {
+    public void setTotalCount(long totalCount) {
         this.totalCount = totalCount;
     }
